fix: Various notation changes  (#287)
* fix(homology): simplexes → simplices

Using "simplexes" isn't wrong, but since "simplices" is used everywhere
else, I'll change it to "simplices" for consistency.

* fix(homology) Hₙ is not only hTop → Grp but also hTop → Ab

* fix(linalg): Use ℝx instead of xℝ

Since the real numbers form a field, if x is a real number, ℝx ≅ xℝ.
However, the constraint that x is real is not mentioned, and the book
defines a vector as a special case of a left R-module, so I think
placing the scalar on the left is a better notation.

* fix(linalg): Let’s not call elements of an R-module vectors.

I know that conventionally many books call elements of an R-module
vectors even when there is no constraint that R is a field. But I
wondered, is that really necessary?